Configured the template for use in the log service
