+++ serverless.yml
@@ -1,54 +1,24 @@
-service: mes-microservices-starter-template
+service: mes-logs-service
 
 provider:
   name: aws
   runtime: nodejs8.10
-  environment:
-    DDB_TABLE_NAME: ddbTable
 
 package:
   individually: true
 
 functions:
   get:
-    handler: src/functions/get/framework/handler.handler
+    handler: src/functions/post/framework/handler.handler
     events:
       - http:
-          path: 'template/{staffNumber}/get'
-          method: get
+          path: 'logs'
+          method: post
 
 custom:
-  dynamodb:
-    start:
-      port: 8000
-      inMemory: true
-      migrate: true
-      seed: true
-    seed:
-      test:
-        sources:
-          - table: ddbTable
-            sources: [./test-data/dynamo-seed-ddbTable.json]
   webpack:
     webpackConfig: 'webpack-sls-offline.config.js'
 
-resources:
-  Resources:
-    penaltyDocuments:
-      Type: AWS::DynamoDB::Table
-      Properties:
-        TableName: ${self:provider.environment.DDB_TABLE_NAME}
-        AttributeDefinitions:
-          - AttributeName: staffNumber
-            AttributeType: S
-        KeySchema:
-          - AttributeName: staffNumber
-            KeyType: HASH
-        ProvisionedThroughput:
-          ReadCapacityUnits: 5
-          WriteCapacityUnits: 5
-
 plugins:
   - serverless-webpack
-  - serverless-dynamodb-local
   - serverless-offline
