--- src/common/domain/README.md
@@ -1,5 +1,0 @@
-# Dynamically generated schema interface goes here
-
-The type information for the schema should live here as `schema.d.ts` - if you don't have this file it can by generated by doing `npm run bootstrap`.
-
-The schema can be found in `/schema`.
